refactor permission spec initialization
